fix Display Columns following pull request comments
