+++ .babelrc
@@ -1,3 +1,3 @@
 {
-    "presets": ["@babel/preset-env", "@babel/preset-react"]
+    "presets": ["@babel/preset-env", "@babel/preset-react", "@babel/preset-flow"]
 }
