+++ src/DisplayColumns/Button/Button.js
@@ -1,0 +1,15 @@
+// @flow
+
+import React from "react";
+import IconButton from "@material-ui/core/IconButton";
+import ViewWeekIcon from "@material-ui/icons/ViewWeek";
+
+const Button = ({ open }: { open: Object => void }) => {
+  return (
+    <IconButton onClick={open}>
+      <ViewWeekIcon />
+    </IconButton>
+  );
+};
+
+export default Button;
