+++ src/DisplayColumns/Displaying/Displaying.js
@@ -1,0 +1,79 @@
+// @flow
+
+import React, { useState, useEffect } from "react";
+import Menu from "@material-ui/core/Menu";
+import useO2xpProvider from "../../hooks/useO2xpProvider";
+import MenuItem from "./MenuItem";
+
+type Props = {
+  anchor: Object,
+  setAnchor: Object => void
+};
+
+const Displaying = ({ anchor, setAnchor }: Props) => {
+  const {
+    state: { data, columns },
+    dispatch
+  } = useO2xpProvider();
+
+  const [allColumns, setAllColumns] = useState([]);
+
+  useEffect(() => {
+    if ((columns.data.columns: Object)) {
+      const newAllColumns: Object[] = Object.keys((columns.data.columns: Object)).map(
+        key => columns.data.columns[key]
+      );
+      setAllColumns(newAllColumns);
+    }
+  }, [columns.data.columns]);
+
+  const handleClose = () => {
+    setAnchor(null);
+  };
+
+  const setColumnVisibility = (column: string) => {
+    const index = allColumns.indexOf(column);
+    let newColumnsOrder: string[] = columns.data.columnsOrder;
+    if (columns.data.columnsOrder.includes(column)) {
+      newColumnsOrder = columns.data.columnsOrder.filter((col: string) => col !== column);
+    } else {
+      newColumnsOrder.splice(index, 0, column);
+    }
+
+    const action = {
+      type: "SET_COLUMNS",
+      payload: {
+        data: {
+          ...columns.data,
+          columnsOrder: newColumnsOrder
+        }
+      }
+    };
+    dispatch(action);
+  };
+
+  return (
+    <div>
+      <Menu
+        id="o2xp-customized-menu"
+        anchorEl={anchor}
+        keepMounted
+        open={Boolean(anchor)}
+        onClose={handleClose}
+        className="o2xp-displayMenu"
+      >
+        {allColumns.map((column: Object) => (
+          <MenuItem
+            className={`o2xp-menu-item-${column.id}`}
+            key={column.id}
+            column={column}
+            columnsOrder={columns.data.columnsOrder}
+            handleClick={setColumnVisibility}
+          />
+        ))}
+      </Menu>
+    </div>
+  );
+};
+
+export default Displaying;
