+++ src/DisplayColumns/Displaying/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Displaying from "./Displaying";
+
+export default Displaying;
